backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
nibabel==5.2.0
numpy==1.26.2
Pillow==10.1.0
python-multipart==0.0.6
